requirements.txt
fastapi==0.129.0
uvicorn==0.40.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.12.5
pydantic-settings==2.13.0
python-jose[cryptography]==3.3.0
bcrypt==5.0.0
python-dotenv==1.2.1
python-multipart==0.0.6
sentence-transformers==5.2.2
torch==2.10.0
numpy==2.3.5
scipy==1.17.0
scikit-learn==1.8.0
cachetools==7.0.1
umap-learn==0.5.11
